% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAnalysis.R
\name{runAnalysis}
\alias{runAnalysis}
\title{Runs the analytic workflow for the Phase2.1SeverityAugPackage project}
\usage{
runAnalysis(
  chartReview_path = "/4ceData/LocalPatientChartReview_wpids.csv",
  maxdays = 30,
  mindays = -2,
  outcome = "Severe?"
)
}
\description{
Runs the analytic workflow for the Phase2.1SeverityAugPackage project
}
\keyword{4CE}
