% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smallCountObfuscation.R
\name{smallCountObfuscation}
\alias{smallCountObfuscation}
\title{Obfuscates small counts in result data for sites where IRBs require such for the Phase2.1SeverityAugPackage project}
\usage{
smallCountObfuscation()
}
\description{
Obfuscates small counts in result data for sites where IRBs require such for the Phase2.1SeverityAugPackage project
}
\keyword{4CE}
